% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lm_plot.R
\name{gg_lm_plot}
\alias{gg_lm_plot}
\title{gg_lm_plot
Code adapated from https://sejohnston.com/2012/08/09/a-quick-and-easy-function-to-plot-lm-results-in-r/}
\usage{
gg_lm_plot(mod, xlab. = NULL, ylab. = NULL, rnd = 2,
  annotate. = "summary")
}
\description{
gg_lm_plot
Code adapated from https://sejohnston.com/2012/08/09/a-quick-and-easy-function-to-plot-lm-results-in-r/
}
