% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammalsmilk-package.R
\docType{package}
\name{mammalsmilk-package}
\alias{mammalsmilk}
\alias{mammalsmilk-package}
\title{mammalsmilk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
