% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_fat.R
\docType{data}
\name{milk_fat}
\alias{milk_fat}
\title{Fat content column from Skiebel et al.}
\format{An object of class \code{data.frame} with 130 rows and 15 columns.}
\usage{
milk_fat
}
\description{
Fat content column from Skiebel et al.
}
\keyword{datasets}
