% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_primates.R
\docType{data}
\name{milk_primates}
\alias{milk_primates}
\title{Just primates}
\format{An object of class \code{data.frame} with 42 rows and 15 columns.}
\usage{
milk_primates
}
\description{
Just primates
}
\keyword{datasets}
