% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_raw.R
\docType{data}
\name{milk_raw}
\alias{milk_raw}
\title{Raw data from Skiebel et al.}
\format{An object of class \code{data.frame} with 130 rows and 22 columns.}
\usage{
milk_raw
}
\description{
Raw data from Skiebel et al.
}
\keyword{datasets}
